+++ modules-services/service-templates/src/main/java/org/openprovenance/prov/service/TemplateLogic.java
@@ -13,6 +13,7 @@ import org.openprovenance.prov.client.RecordsProcessorInterface;
 import org.openprovenance.prov.model.QualifiedName;
 import org.openprovenance.prov.service.core.ServiceUtils;
 import org.openprovenance.prov.service.iobean.composite.SqlCompositeBeanEnactor3;
+import org.openprovenance.prov.service.readers.TemplatesVizConfig;
 import org.openprovenance.prov.template.library.plead.Plead_trainingBuilder;
 import org.openprovenance.prov.template.log2prov.FileBuilder;
 import org.openprovenance.prov.vanilla.ProvFactory;
@@ -149,7 +150,7 @@ public class TemplateLogic {
     }
 
 
-    public void generateViz(TemplateService.TemplatesVizConfig config, OutputStream out) {
+    public void generateViz(TemplatesVizConfig config, OutputStream out) {
 
         typeAssignment.entrySet().removeIf(entry -> entry.getValue() ==null || entry.getValue().isEmpty());