+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/sql/QueryBuilder.java
@@ -499,6 +499,16 @@ public class QueryBuilder {
         return this;
 
     }
+    public QueryBuilder allowBreak() {
+        prettyPrinter.allowBreak(0);
+        return this;
+    }
+    public QueryBuilder begin(int n) {
+        prettyPrinter.begin(n);
+        return this;
+    }
+
+
 
     private QueryBuilder conditionalStart(String operator) {
         prettyPrinter.write(operator);